% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality.R
\name{check_normality}
\alias{check_normality}
\title{Check Normality and Summarize Data}
\usage{
check_normality(data, variable)
}
\arguments{
\item{data}{A dataframe containing the data.}

\item{variable}{A string specifying the column name of the variable to be checked and summarized.}
}
\value{
A list containing the summary statistics (mean and standard deviation if normal, median and IQR if not normal).
}
\description{
This function checks the normality of a specified variable in a dataframe using the Shapiro-Wilk test
and provides summary statistics (mean and standard deviation if normal, median and IQR if not normal).
}
\examples{
# Example usage with a dataframe 'df' and outcome variable 'business_days_until_appointment'
check_normality(df, "business_days_until_appointment")

}
