% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_section_logistic_regression.R
\name{logistic_regression}
\alias{logistic_regression}
\title{Perform Logistic Regression on Multiple Predictors with Logging}
\usage{
logistic_regression(df, target_variable, predictor_vars, significance_level)
}
\arguments{
\item{df}{A data frame containing the target and predictor variables.}

\item{target_variable}{A string representing the name of the target variable.}

\item{predictor_vars}{A vector of strings representing the names of predictor variables.}

\item{significance_level}{A numeric value specifying the significance level for filtering predictors.}
}
\value{
A data frame containing significant predictors with their p-values and formatted p-values.
}
\description{
This function performs logistic regression for multiple predictor variables
against a specified target variable. It logs the process, including inputs,
data transformations, and results. The function filters predictors based on
a given significance level.
}
\examples{
# Assuming df is your data frame
target_variable <- "cleaned_does_the_physician_accept_medicaid_numeric"
predictor_vars <- setdiff(names(df), c(target_variable, "does_the_physician_accept_medicaid", "cleaned_does_the_physician_accept_medicaid"))
significance_logistic_regression <- 0.2

significant_vars <- logistic_regression(df, target_variable, predictor_vars, significance_level = significance_logistic_regression)
print(significant_vars)
}
