% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nppes_collect_and_clean_data.R
\name{nppes_collect_and_clean_data}
\alias{nppes_collect_and_clean_data}
\title{Collect and clean the processed data}
\usage{
nppes_collect_and_clean_data(processed_data)
}
\arguments{
\item{processed_data}{A \code{tbl} object containing the processed data.}
}
\value{
A cleaned data frame.
}
\description{
This function collects and cleans the processed data from DuckDB.
}
\details{
After collecting and cleaning the data, the next step is to save the cleaned data to a CSV file using \code{nppes_save_data_to_csv()}.
}
\examples{
cleaned_data <- nppes_collect_and_clean_data(processed_data)
# Next step: nppes_save_data_to_csv()
}
