% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nppes_connect_to_duckdb.R
\name{nppes_connect_to_duckdb}
\alias{nppes_connect_to_duckdb}
\title{Connect to DuckDB}
\usage{
nppes_connect_to_duckdb(duckdb_file_path)
}
\arguments{
\item{duckdb_file_path}{The file path to the DuckDB database file.}
}
\value{
A DuckDB connection object.
}
\description{
This function establishes a connection to a DuckDB database.
}
\examples{
con <- nppes_connect_to_duckdb("/path/to/your/duckdb_file.duckdb")
}
