% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nppes_create_table_from_csv.R
\name{nppes_create_table_from_csv}
\alias{nppes_create_table_from_csv}
\title{Create a table in DuckDB from a CSV file}
\usage{
nppes_create_table_from_csv(con, file_path, table_name = "npi_2024")
}
\arguments{
\item{con}{A DuckDB connection object.}

\item{file_path}{A character string specifying the path to the CSV file.}

\item{table_name}{A character string specifying the name of the table to be created. Default is "npi_2024".}
}
\value{
None
}
\description{
This function creates a table in DuckDB by reading data from a CSV file.
}
\details{
After creating the table, the next step is to query sample data using \code{nppes_query_sample_data()}.
}
\examples{
con <- nppes_connect_to_duckdb("/path/to/your/duckdb_file.duckdb")
nppes_create_table_from_csv(con, "/path/to/your/csv_file.csv", "npi_2024")
# Next step: nppes_query_sample_data()

con <- nppes_connect_to_duckdb("/Volumes/Video Projects Muffly 1/nppes_historical_downloads/my_duckdb.duckdb")
nppes_create_table_from_csv(con, "/Volumes/Video Projects Muffly 1/nppes_historical_downloads/NPPES_Data.csv", "npi_2024")
# Next step: nppes_query_sample_data()
}
