% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nppes_disconnect_from_duckdb.R
\name{nppes_disconnect_from_duckdb}
\alias{nppes_disconnect_from_duckdb}
\title{Disconnect from DuckDB}
\usage{
nppes_disconnect_from_duckdb(con)
}
\arguments{
\item{con}{A DuckDB connection object.}
}
\value{
None
}
\description{
This function disconnects from a DuckDB database connection after all operations are complete.
}
\details{
Use this function after you have completed all operations involving the DuckDB connection.
}
\examples{
con <- nppes_connect_to_duckdb("/path/to/your/duckdb_file.duckdb")
# After all operations, disconnect from the database
nppes_disconnect_from_duckdb(con)

con <- nppes_connect_to_duckdb("/Volumes/Video Projects Muffly 1/nppes_historical_downloads/my_duckdb.duckdb")
# After all operations, disconnect from the database
nppes_disconnect_from_duckdb(con)
}
