% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_phase_1_results.R,
%   R/get_nppes_data_for_one_year.R
\name{nppes_get_data_for_one_year}
\alias{nppes_get_data_for_one_year}
\title{Process NPPES Data for One Year with Chunked Processing}
\usage{
nppes_get_data_for_one_year(
  npi_file_path,
  output_csv_path,
  duckdb_file_path =
    "/Volumes/Video Projects Muffly 1/nppes_historical_downloads/my_duckdb.duckdb",
  taxonomy_codes_1 = c("207V00000X", "207VB0002X", "207VC0300X", "207VC0200X",
    "207VX0201X", "207VG0400X", "207VH0002X", "207VM0101X", "207VX0000X", "207VE0102X",
    "207VF0040X"),
  taxonomy_codes_2 = c("207V00000X", "207VB0002X", "207VC0300X", "207VC0200X",
    "207VX0201X", "207VG0400X", "207VH0002X", "207VM0101X", "207VX0000X", "207VE0102X",
    "207VF0040X"),
  save_column_in_each_nppes_year = FALSE,
  excel_file_path = NULL
)

nppes_get_data_for_one_year(
  npi_file_path,
  output_csv_path,
  duckdb_file_path =
    "/Volumes/Video Projects Muffly 1/nppes_historical_downloads/my_duckdb.duckdb",
  taxonomy_codes_1 = c("207V00000X", "207VB0002X", "207VC0300X", "207VC0200X",
    "207VX0201X", "207VG0400X", "207VH0002X", "207VM0101X", "207VX0000X", "207VE0102X",
    "207VF0040X"),
  taxonomy_codes_2 = c("207V00000X", "207VB0002X", "207VC0300X", "207VC0200X",
    "207VX0201X", "207VG0400X", "207VH0002X", "207VM0101X", "207VX0000X", "207VE0102X",
    "207VF0040X"),
  save_column_in_each_nppes_year = FALSE,
  excel_file_path = NULL
)
}
\arguments{
\item{npi_file_path}{A character string specifying the path to the raw NPPES data CSV file.}

\item{output_csv_path}{A character string specifying the full file path where the cleaned data will be saved.}

\item{duckdb_file_path}{A character string specifying the path to the DuckDB database file.}

\item{taxonomy_codes_1}{A character vector of taxonomy codes used to filter data in \verb{Healthcare Provider Taxonomy Code_1}.}

\item{taxonomy_codes_2}{A character vector of taxonomy codes used to filter data in \verb{Healthcare Provider Taxonomy Code_2}.}

\item{save_column_in_each_nppes_year}{A logical value indicating whether to save a sample to an Excel file.}

\item{excel_file_path}{A character string specifying the path to save the Excel file if \code{save_column_in_each_nppes_year} is \code{TRUE}.}
}
\value{
A data frame containing the cleaned NPPES data for one year, saved to the specified file path.

A data frame containing the cleaned NPPES data for one year, saved to the specified file path.
}
\description{
Process NPPES Data for One Year with Chunked Processing

Process NPPES Data for One Year with Chunked Processing
}
