% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nppes_initialize_environment.R
\name{nppes_initialize_environment}
\alias{nppes_initialize_environment}
\title{Initialize the environment by loading necessary libraries and resolving conflicts}
\usage{
nppes_initialize_environment(setup_file)
}
\value{
None
}
\description{
This function sets up the environment by sourcing the setup script and resolving package conflicts.
}
\details{
After initializing the environment, the next step is to connect to DuckDB using \code{connect_to_duckdb()}.
}
\examples{
nppes_initialize_environment()
# Next step: tyler::connect_to_duckdb()
}
