% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nppes_process_npi_table.R
\name{nppes_process_npi_table_chunk}
\alias{nppes_process_npi_table_chunk}
\title{Process an NPI table in DuckDB}
\usage{
nppes_process_npi_table_chunk(
  con,
  table_name = "npi_2024",
  taxonomy_codes_1 = c("390200000X", "174400000X", "207V00000X", "207VB0002X",
    "207VC0300X", "207VC0200X", "207VX0201X", "207VG0400X", "207VH0002X", "207VM0101X",
    "207VX0000X", "207VE0102X", "207VF0040X"),
  taxonomy_codes_2 = c("390200000X", "174400000X", "207V00000X", "207VB0002X",
    "207VC0300X", "207VC0200X", "207VX0201X", "207VG0400X", "207VH0002X", "207VM0101X",
    "207VX0000X", "207VE0102X", "207VF0040X")
)
}
\arguments{
\item{con}{A DuckDB connection object.}

\item{table_name}{A character string specifying the name of the table to process. Default is "npi_2024".}

\item{taxonomy_codes_1}{A character vector specifying the taxonomy codes to filter in \verb{Healthcare Provider Taxonomy Code_1}. Default is a predefined set of codes.}

\item{taxonomy_codes_2}{A character vector specifying the taxonomy codes to filter in \verb{Healthcare Provider Taxonomy Code_2}. Default is a predefined set of codes.}
}
\value{
A \code{tbl} object containing the processed data.
}
\description{
This function processes an NPI table in DuckDB, filtering by taxonomy codes and cleaning the data.
}
\details{
After processing the NPI table, the next step is to collect and clean the data using \code{nppes_collect_and_clean_data()}.
}
\examples{
con <- nppes_connect_to_duckdb("/path/to/your/duckdb_file.duckdb")
processed_data <- nppes_process_npi_table_chunk(con, "npi_2024", c("207V00000X"), c("207V00000X"))
# Next step: nppes_collect_and_clean_data()

con <- nppes_connect_to_duckdb("/Volumes/Video Projects Muffly 1/nppes_historical_downloads/my_duckdb.duckdb")
processed_data <- nppes_process_npi_table_chunk(con)
# Next step: nppes_collect_and_clean_data()
}
