% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nppes_query_sample_data.R
\name{nppes_query_sample_data}
\alias{nppes_query_sample_data}
\title{Query sample data from a DuckDB table}
\usage{
nppes_query_sample_data(
  con,
  table_name = "npi_2024",
  limit = 5,
  save_column_in_each_nppes_year = TRUE,
  excel_file_path = NULL
)
}
\arguments{
\item{con}{A DuckDB connection object.}

\item{table_name}{A character string specifying the name of the table to query. Default is "npi_2024".}

\item{limit}{An integer specifying the number of rows to retrieve. Default is 5.}

\item{save_column_in_each_nppes_year}{A logical value indicating whether to save the sample data to an Excel file. Default is FALSE.}

\item{excel_file_path}{A character string specifying the path to the Excel file if \code{save_column_in_each_nppes_year} is TRUE. Default is NULL.}
}
\value{
A data frame containing the queried sample data.
}
\description{
This function queries a sample of data from a specified DuckDB table.
}
\details{
After querying the sample data, the next step is to process the NPI table using \code{nppes_process_npi_table_chunk()}.
}
\examples{
con <- nppes_connect_to_duckdb("/path/to/your/duckdb_file.duckdb")
nppes_query_sample_data(con, "npi_2024", 5)
# Next step: nppes_process_npi_table_chunk()

con <- nppes_connect_to_duckdb("/Volumes/Video Projects Muffly 1/nppes_historical_downloads/my_duckdb.duckdb")
nppes_query_sample_data(con, "npi_2024", 10, TRUE, "/path/to/excel_file.xlsx")
# Next step: nppes_process_npi_table_chunk()
}
