% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nppes_save_data_to_csv.R
\name{nppes_save_data_to_csv}
\alias{nppes_save_data_to_csv}
\title{Save the final data to CSV}
\usage{
nppes_save_data_to_csv(data, file_path)
}
\arguments{
\item{data}{A cleaned data frame.}

\item{file_path}{A character string specifying the path to the CSV file.}
}
\value{
None
}
\description{
This function saves the cleaned data to a CSV file.
}
\details{
Use this function to save the final cleaned data to a specified file path.
}
\examples{
nppes_save_data_to_csv(cleaned_data, "/path/to/your/file.csv")
}
