% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{prepare_dataset}
\alias{prepare_dataset}
\title{Prepare the Dataset by Excluding Certain Columns with Logging}
\usage{
prepare_dataset(df, target_variable, excluded_columns)
}
\arguments{
\item{df}{A data frame containing the dataset.}

\item{target_variable}{A string representing the name of the target variable.}

\item{excluded_columns}{A vector of strings representing the names of columns to exclude from the predictors.}
}
\value{
A vector of strings representing the names of predictor variables.
}
\description{
This function prepares the dataset by excluding specified columns from the predictor variables.
It logs the process, including inputs and outputs.
}
