% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_section_most_common_provider_info.r
\name{results_section_summarize_common_provider_info}
\alias{results_section_summarize_common_provider_info}
\title{Summarize the Most Common Gender, Specialty, Training, and Academic Affiliation in Provider Data}
\usage{
results_section_summarize_common_provider_info(
  provider_info,
  gender_col,
  specialty_col,
  training_col,
  academic_affiliation_col
)
}
\arguments{
\item{provider_info}{A data frame containing provider information.}

\item{gender_col}{Name of the column representing gender.}

\item{specialty_col}{Name of the column representing specialty.}

\item{training_col}{Name of the column representing training credentials.}

\item{academic_affiliation_col}{Name of the column representing academic affiliation.}
}
\value{
A character string summarizing the most common gender, specialty, training, and academic affiliation along with their respective proportions.
}
\description{
This function calculates and returns a summary sentence describing the most common gender, specialty, training, and academic affiliation in the provided dataset.
It filters out missing values in each column before determining the most common value, then calculates the proportion of this most common value relative to the total non-missing values in that column.
}
\examples{
# Example usage with specified columns
provider_info <- data.frame(
  gender = c("Male", "Female", "Female", "Male", "Male"),
  specialty = c("Cardiology", "Cardiology", "Neurology", "Cardiology", "Neurology"),
  Provider.Credential.Text = c("MD", "MD", "DO", "MD", "DO"),
  academic_affiliation = c("Yes", "No", "Yes", "No", "Yes")
)
summarize_common_provider_info(
  provider_info,
  gender_col = "gender",
  specialty_col = "specialty",
  training_col = "Provider.Credential.Text",
  academic_affiliation_col = "academic_affiliation"
)

}
