% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_and_process_npi.R
\name{search_and_process_npi}
\alias{search_and_process_npi}
\title{Search and Process NPI Numbers with Logging}
\usage{
search_and_process_npi(
  name_data,
  npi_type = "ind",
  max_results_limit = 5L,
  valid_country = "US",
  allowed_credentials = c("MD", "DO"),
  searchable_taxonomies = c("Allergy & Immunology", "Anesthesiology", "Dermatology",
    "Family Medicine", "Internal Medicine", "Obstetrics & Gynecology", "Pediatrics",
    "Psychiatry & Neurology", "Radiology", "Surgery", "Urology"),
  chunk_size = 10,
  output_directory = NULL
)
}
\arguments{
\item{name_data}{A data frame with columns 'first' and 'last' containing the names to search.}

\item{npi_type}{The enumeration type for NPI search (e.g., "ind", "org", "all"). Default is "ind".}

\item{max_results_limit}{The maximum number of search results to request for each name pair. Default is 5.}

\item{valid_country}{Filter for only valid country, default is "US".}

\item{allowed_credentials}{A character vector containing the credentials to filter the NPI results. Default is c("MD", "DO").}

\item{searchable_taxonomies}{A character vector of taxonomies to filter NPI results. Default is a set of clinical taxonomies.  A searchable list can be found at the Medicare Provider and Supplier Taxonomy Crosswalk: https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-provider-and-supplier-taxonomy-crosswalk/data.}

\item{chunk_size}{The number of results to save per chunk. Default is 10.}

\item{output_directory}{Directory to save chunked results. Default is NULL (current working directory).}
}
\value{
A combined data frame of all NPI search results.
}
\description{
This function searches for NPI numbers based on provided first and last names, filters results by taxonomy,
and writes chunked results to CSV files. It supports retrying failed searches and ensures logging of inputs, outputs,
and errors.
}
