% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase0_search_by_taxonomy.R
\name{search_by_taxonomy}
\alias{search_by_taxonomy}
\title{Search NPI by Taxonomy}
\usage{
search_by_taxonomy(taxonomy_to_search)
}
\arguments{
\item{taxonomy_to_search}{A character vector of taxonomy descriptions to search for in the NPI registry.}
}
\value{
A dataframe of cleaned NPI data, filtered by taxonomy and other criteria.
}
\description{
This function performs a search on the NPI registry for a list of taxonomy descriptions and returns a combined dataframe of results.
It processes the returned data and cleans up unnecessary fields.
}
