% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_section_generate_overall_table.R
\name{tm_write2pdf}
\alias{tm_write2pdf}
\title{Save Arsenal Table to PDF}
\usage{
tm_write2pdf(object, filename)
}
\arguments{
\item{object}{An Arsenal table object.}

\item{filename}{The file path for saving the PDF.}
}
\description{
This function saves an Arsenal table as a PDF.
}
