% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tyler-package.R
\docType{package}
\name{tyler-package}
\alias{tyler}
\alias{tyler-package}
\title{tyler: Common Functions for Mystery Caller or Audit Studies Evaluating Patient Access to Care}
\description{
The 'tyler' package provides a collection of functions designed to facilitate mystery caller studies, often used in evaluating patient access to healthcare. It includes tools for searching and processing National Provider Identifier (NPI) numbers based on names and analyzing demographic data associated with these NPIs. The package simplifies the handling of NPI data and the creation of informative tables for analysis and reporting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mufflyt.github.io/mystery/}
  \item Report bugs at \url{https://github.com/mufflyt/tyler/issues}
}

}
\keyword{NPI}
\keyword{NPPES}
\keyword{healthcare}
