% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_create_region_map.R
\name{create_region_map}
\alias{create_region_map}
\title{Main function to create a map of U.S. States by Region, Division, or
Custom Districts}
\usage{
create_region_map(
  remove_ak_hi = TRUE,
  districts_per_group = "acog_districts",
  save_path = NULL,
  alpha_level = 0.4,
  title = "U.S. States by Region/Division or Custom Districts",
  subtitle = NULL
)
}
\arguments{
\item{remove_ak_hi}{Logical, whether to exclude Alaska and Hawaii from the
map. Default is TRUE.}

\item{districts_per_group}{Character, which grouping to use for the regions.
Options are:
"acog_districts", "ENT_Board_of_Governors_Regions", or
"US_Census_Subdivisions".}

\item{save_path}{Character, an optional file path to save the map image.
If NULL, the map will not
be saved. Default is NULL.}

\item{alpha_level}{Numeric, the transparency level for the map's fill color,
with 0 being fully
transparent and 1 fully opaque. Default is 0.2.}

\item{title}{Character, the title for the map. Default is "U.S. States by
Region/Division or Custom
Districts".}

\item{subtitle}{Character, an optional subtitle for the map.}
}
\value{
A ggplot map object representing the U.S. states colored by
region/division, ACOG districts,
or ENT_Board_of_Governors_Regions.
}
\description{
This function generates a map of U.S. states, coloring them by regions,
divisions, ACOG districts,
or ENT Board of Governors Regions. Users can customize which grouping to
use (ACOG Districts,
ENT_Board_of_Governors_Regions, or US Census Subdivisions).
}
\examples{
# Example 1: Create a map of U.S. states by region/division,
# Excluding Alaska
data(us_census_bureau_regions_df)
regions_map <- us_census_bureau_regions_df \%>\%
  dplyr::filter(State != "Alaska") \%>\%
  ggplot2::ggplot(ggplot2::aes(fill = Region)) +
  ggplot2::geom_map(map = us_map, ggplot2::aes(map_id = State.Code)) +
  ggplot2::labs(title = "Map of U.S. Regions (Excluding Alaska)")
print(regions_map)

# Example 2: Create a map of U.S. states by ACOG districts,
including Alaska and Hawaii
create_region_map(
  remove_ak_hi = FALSE,
  districts_per_group = "acog_districts",
  alpha_level = 0.5
)

# Example 3: Create a map of U.S. states by ENT Board of Governors Regions
and save it to a file
create_region_map(
  remove_ak_hi = TRUE,
  districts_per_group = "ENT_Board_of_Governors_Regions",
  save_path = "ent_bog_regions_map.png",
  alpha_level = 0.3
)

}
