% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_physicians_by_group.R
\name{map_physicians_by_group}
\alias{map_physicians_by_group}
\title{Map Physicians by State or Subdivision}
\usage{
map_physicians_by_group(
  state_counts,
  output_file_prefix = "choropleth_map",
  group_by = "state"
)
}
\arguments{
\item{state_counts}{A tibble containing the counts of physicians per state or subdivision. Must include columns \code{region} and \code{count}.}

\item{output_file_prefix}{A character string specifying the prefix for the output PNG file of the choropleth map. Default is "choropleth_map".}

\item{group_by}{A character string indicating whether the counts are grouped by "state" or "subdivision". Default is "state".}
}
\value{
A ggplot2 object representing the choropleth map.
}
\description{
This function generates a choropleth map based on physician counts and saves the map as a PNG file.
}
\examples{
# Example 1: Map physicians by state
\dontrun{
state_counts <- count_physicians_by_group(taxonomy_and_aaos_data)
map_physicians_by_group(state_counts)
}

# Example 2: Map physicians by U.S. Census Bureau subdivision
\dontrun{
subdivision_counts <- count_physicians_by_group(taxonomy_and_aaos_data, group_by = "subdivision")
map_physicians_by_group(subdivision_counts, group_by = "subdivision")
}

# Example 3: Save the map with a custom output file prefix
\dontrun{
state_counts <- count_physicians_by_group(taxonomy_and_aaos_data)
map_physicians_by_group(state_counts, output_file_prefix = "custom_map_prefix")
}

}
